Изменена версия либы. Ошибки вынесены в отдельный файл
