+++ bank_account_parser/src/lib.rs
@@ -4,3 +4,4 @@ mod common;
 pub mod csv_format;
 pub mod mt940_format;
 pub mod transactions_holder;
+mod error;
